DEV: adapt the data type from long to uint type

+++ cs-trinity/BlockChain/NeoInterface.cs
@@ -398,7 +398,7 @@ namespace Trinity.BlockChain
             return BigInteger.Parse(i.ToString()).ToByteArray().ToHexString();
         }
 
-        public static string BlockheightToScript(long input)
+        public static string BlockheightToScript(uint input)
         {
             string inputHex = Convert.ToString(input, 16);
